refactored my projects into helpers
